Don't cache the metadata lookup promise.
Caching this promise is problematic when the promise rejects for any reason (no available connections in DB pool or any other error running the query). Essentially the rejected promise will be returned every time this function is called for the same table. I described this in a little more detail and linked to a repro case in https://github.com/Vincit/objection.js/issues/1997

This fix is incomplete:
- Now that we're no longer caching the promise we should clean up the code. There are a few places that check "is the cached item a promise or is it real metadata?" There's no longer any need to check if it's a promise.
- There are no tests.
- The biggest downside of this commit is that it removes the benefit of caching the promise: If `fetchTableMetadata()` is called many times in quick succession then it's nice to reuse the promise rather than looking up the metadata many times.

A better fix would be to continue caching the promise but unset it if the promise rejects. That can be done by appending this to the metadata query promise chain:
```
.catch((err) => {
  modelClass[TABLE_METADATA].set(table, undefined);
  throw err;
})
```
But that doesn't fix the problem in all cases: When `fetchTableMetadata()` is called from RelationJoiner it uses `mapPromise` to evaluate a bunch of metadata promises at once. `mapPromise` rejects as soon as a single promise rejects. It might be sufficient to change RelationJoined to use used `Promise.allSettled()`.