Remove rejected metadata promises from the cache.
These promises can reject (no available connections in DB pool, any other error running the query, etc). It's important to remove the rejected promise from the cache so that `fetchTableMetadata()` doesn't keep returning it. I described this in a little more detail and linked to a repro case in https://github.com/Vincit/objection.js/issues/1997

This fix is incomplete:
- There are no tests.
- See the TODO about needing to use `Promise.allSettled()`.